refactor: move all covid_symptom code into its own submodule
This isolates the study-specific code and gives an example for future
studies / further covid code.
