+++ cumulus_etl/etl/studies/covid_symptom/__init__.py
@@ -1,0 +1,3 @@
+"""The covid_symptom study"""
+
+from .covid_tasks import CovidSymptomNlpResultsTask
