+++ cumulus_etl/nlp/__init__.py
@@ -1,4 +1,4 @@
 """Support code for NLP servers"""
 
-from .extract import covid_symptoms_extract, ctakes_httpx_client
+from .extract import ctakes_extract, ctakes_httpx_client, list_polarity
 from .watcher import check_cnlpt, check_ctakes, restart_ctakes_with_bsv
